pyproject.toml
[build-system]
requires = [
    "scikit_build_core",
    "pybind11",
    "numpy",
]
build-backend = "scikit_build_core.build"

[project]
name = "alphafold3"
version = "3.0.1"
requires-python = ">=3.12"
readme = "README.md"
license = {file = "LICENSE"}
dependencies = [
    "absl-py>=2.3.1",
    "dm-haiku==0.0.16",
    "dm-tree",
    "h5py",
    "jax==0.9.1",
    "jax[cuda12]==0.9.1",
    "numpy",
    "rdkit",
    "tokamax==0.0.11",
    "tqdm",
    "zstandard",
]

[project.optional-dependencies]
test = ["pytest>=6.0"]

[tool.scikit-build]
wheel.exclude = [
    "**.pyx",
    "**/CMakeLists.txt",
    "**.cc",
    "**.h"
]
sdist.include = [
    "LICENSE",
    "OUTPUT_TERMS_OF_USE.md",
    "WEIGHTS_PROHIBITED_USE_POLICY.md",
    "WEIGHTS_TERMS_OF_USE.md",
]

[tool.cibuildwheel]
build = "cp3*-manylinux_x86_64"
manylinux-x86_64-image = "manylinux_2_28"

[project.scripts]
build_data = "alphafold3.build_data:build_data"
